Update twitter filter to #io16
This change will also remove retweets: there are just too many of them
and the risk is we'll miss tweets originated from googledevs.

The code could be improved by looping over multiple requests in a task
queue but it adds a lot of complexity. Will get back to this as a P2/P3.